Refactor dynamic rendering and import paths
