+++ apps/earn-protocol-landing-page/app/server-handlers/mixpanel/index.ts
