+++ apps/earn-protocol-landing-page/components/organisms/LandingPageStrategyPicker/LandingPageStrategyPicker.tsx
@@ -1,11 +1,11 @@
 'use client'
 
 import { StrategyCard, StrategySimulationForm } from '@summerfi/app-earn-ui'
-import { type SDKVaultType } from '@summerfi/app-types'
+import { type SDKVaultishType } from '@summerfi/app-types'
 
 import classNames from '@/components/organisms/LandingPageStrategyPicker/LandingPageStrategyPicker.module.scss'
 
-export const LandingPageStrategyPicker = ({ strategy }: { strategy: SDKVaultType }) => {
+export const LandingPageStrategyPicker = ({ strategy }: { strategy: SDKVaultishType }) => {
   return (
     <div className={classNames.wrapper}>
       <StrategyCard {...strategy} />
