+++ apps/earn-protocol/app/earn/[network]/details/[vaultId]/page.tsx
@@ -1,8 +1,8 @@
 import { StrategyGridDetails, subgraphNetworkToId } from '@summerfi/app-earn-ui'
 import { type SDKNetwork } from '@summerfi/app-types'
 
+import { getVaultDetails } from '@/app/server-handlers/sdk/getVaultDetails'
 import { StrategyDetailsView } from '@/components/layout/StrategyDetailsView/StrategyDetailsView'
-import { getVaultDetails } from '@/server-handlers/sdk/getVaultDetails'
 
 type EarnStrategyDetailsPageProps = {
   params: {
@@ -10,6 +10,8 @@ type EarnStrategyDetailsPageProps = {
     vaultId: string
   }
 }
+
+export const dynamic = 'force-dynamic'
 
 const EarnStrategyDetailsPage = async ({ params }: EarnStrategyDetailsPageProps) => {
   const networkId = subgraphNetworkToId(params.network)
