+++ apps/earn-protocol/app/page.tsx
@@ -1,6 +1,8 @@
 import { Text } from '@summerfi/app-earn-ui'
 import Link from 'next/link'
 
+export const dynamic = 'force-dynamic'
+
 export default function HomePage() {
   return (
     <div
