+++ apps/earn-protocol/features/portfolio/components/PortfolioRewards/PortfolioRewards.tsx
@@ -1,9 +1,9 @@
 import { type FC } from 'react'
 
+import type { PortfolioRewardsRawData } from '@/app/server-handlers/portfolio/portfolio-rewards-handler'
 import { PortfolioRewardsBoost } from '@/features/portfolio/components/PortfolioRewardsBoost/PortfolioRewardsBoost'
 import { PortfolioRewardsMore } from '@/features/portfolio/components/PortfolioRewardsMore/PortfolioRewardsMore'
 import { PortfolioRewardsWhat } from '@/features/portfolio/components/PortfolioRewardsWhat/PortfolioRewardsWhat'
-import type { PortfolioRewardsRawData } from '@/server-handlers/portfolio/portfolio-rewards-handler'
 
 import classNames from './PortfolioRewards.module.scss'
