--- apps/earn-protocol/server-handlers/sdk/constants/sdk-supported-networks.ts
@@ -1,3 +1,0 @@
-import { ChainId } from '@summerfi/serverless-shared'
-
-export const sdkSupportedNetworks = [ChainId.ARBITRUM, ChainId.BASE]
