+++ packages/app-earn-ui/src/components/atoms/Timeframes/Timeframes.module.scss
@@ -1,6 +1,6 @@
 .timeframesWrapper {
   display: flex;
-  justify-content: end;
+  justify-content: flex-end;
   align-items: center;
 
   button {
